backend/requirements.txt
Flask
flask-cors
requests
tldextract
flask
flask-cors
python-dotenv
google-generativeai
